refactor(graph_view): extract graph-geometry module (#76)
Pure code move: GRAPH_LEADING_COLUMNS, the AVATAR_* consts, avatars_active,
avatar_overlay_x, effective_graph_width, next_graph_cap, graph_truncation,
pixel_row_cells, and ellipsis_style (plus their unit tests) move into a new
geometry submodule. The public consts/fns are re-exported so the
graph_view::* contract with ui/mod is preserved exactly. No behavior changes.
